- Display message on login screen if user's session timed out - Add logic to force a logout if app has timed out while in the background

+++ iOS/TeleMedApplication.h
@@ -15,7 +15,7 @@
 	NSTimer *_idleTimer;
 }
 
-@property (nonatomic) int timeoutPeriodMins;
+@property (nonatomic) NSInteger timeoutPeriodMins;
 
 /**
  * Resets the idle timer to its initial state. This method gets called
